requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-environ==0.4.5
django-rest-auth==0.9.5
djangorestframework==3.11.0
Pillow==7.2.0
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.2.0
